+++ apps/meteor/client/components/WarningModal.spec.tsx
@@ -4,8 +4,6 @@ import React from 'react';
 
 import WarningModal from './WarningModal';
 
-import '@testing-library/jest-dom';
-
 it('should look good', async () => {
 	render(<WarningModal text='text' confirmText='confirm' cancelText='cancel' confirm={() => undefined} close={() => undefined} />, {
 		wrapper: mockAppRoot().build(),
