+++ apps/meteor/client/lib/download.spec.ts
@@ -1,47 +1,44 @@
-import { expect, spy } from 'chai';
-import { describe, it } from 'mocha';
-
-import { download, downloadAs, downloadCsvAs, downloadJsonAs } from '../../../../client/lib/download';
+import { download, downloadAs, downloadCsvAs, downloadJsonAs } from './download';
 
 describe('download', () => {
 	it('should work', () => {
-		const listener = spy();
+		const listener = jest.fn();
 		document.addEventListener('click', listener, false);
 
 		download('about:blank', 'blank');
 
 		document.removeEventListener('click', listener, false);
-		expect(listener).to.have.been.called();
+		expect(listener).toHaveBeenCalled();
 	});
 });
 
 describe('downloadAs', () => {
 	it('should work', () => {
-		const listener = spy();
+		const listener = jest.fn();
 		document.addEventListener('click', listener, false);
 
 		downloadAs({ data: [] }, 'blank');
 
 		document.removeEventListener('click', listener, false);
-		expect(listener).to.have.been.called();
+		expect(listener).toHaveBeenCalled();
 	});
 });
 
 describe('downloadJsonAs', () => {
 	it('should work', () => {
-		const listener = spy();
+		const listener = jest.fn();
 		document.addEventListener('click', listener, false);
 
 		downloadJsonAs({}, 'blank');
 
 		document.removeEventListener('click', listener, false);
-		expect(listener).to.have.been.called();
+		expect(listener).toHaveBeenCalled();
 	});
 });
 
 describe('downloadCsvAs', () => {
 	it('should work', () => {
-		const listener = spy();
+		const listener = jest.fn();
 		document.addEventListener('click', listener, false);
 
 		downloadCsvAs(
@@ -53,6 +50,6 @@ describe('downloadCsvAs', () => {
 		);
 
 		document.removeEventListener('click', listener, false);
-		expect(listener).to.have.been.called();
+		expect(listener).toHaveBeenCalled();
 	});
 });
