+++ apps/meteor/client/views/marketplace/AppDetailsPage/tabs/AppStatus/AppStatus.spec.tsx
@@ -1,8 +1,8 @@
 import { faker } from '@faker-js/faker';
 import { mockAppRoot } from '@rocket.chat/mock-providers';
 import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import React from 'react';
-import '@testing-library/jest-dom';
 
 import { mockedAppsContext } from '../../../../../../tests/mocks/client/marketplace';
 import { createFakeApp, createFakeLicenseInfo } from '../../../../../../tests/mocks/data';
@@ -28,5 +28,5 @@ it('should look good', async () => {
 			.build(),
 	});
 
-	screen.getByRole('button', { name: 'Request' }).click();
+	userEvent.click(screen.getByRole('button', { name: 'Request' }));
 });
