+++ apps/meteor/client/views/marketplace/AppMenu.spec.tsx
@@ -2,7 +2,6 @@ import { faker } from '@faker-js/faker';
 import { mockAppRoot } from '@rocket.chat/mock-providers';
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-import '@testing-library/jest-dom';
 
 import { mockedAppsContext } from '../../../tests/mocks/client/marketplace';
 import { createFakeApp } from '../../../tests/mocks/data';
